OneDrive/Desktop/chutiyaap/movie-recommender/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
requests
python-dotenv
